requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-classic
langchain-text-splitters
faiss-cpu
pypdf
tiktoken
python-dotenv
openai
pytest
